fix: remove requirement for babel-polyfill
If you have babel-polyfill in your own app and register it, we’d run into
an error where babel complains that we required the polyfill multiple times.
Since we only need Promises polyfilled, we just use bluebird for the case
where Promises are not available in the runtime. This also has the upside of
removing babel from the regular dependencies, which should reduce install times.